backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
mysql-connector-python==8.0.33
python-dotenv==0.21.1
pandas==1.5.3
chardet==5.1.0
python-multipart==0.0.6
typing-extensions==4.5.0
numpy==1.24.3
pydantic==1.10.7
